# api/openapi.yaml
openapi: 3.1.0
info:
  title: Gachify API
  version: 0.6.0
  description: |
    HTTP API for the Gachify streaming platform.

    - Public routes under `/api/v1`
    - Admin moderation under `/internal/admin` (header `X-Gachify-Admin-Key`)
    - Seed/bootstrap under `/internal/seed` (dev or `X-Gachify-Seed-Key`)
servers:
  - url: http://localhost:8080
    description: Local API
  - url: https://example.com
    description: Production (replace with GACHIFY_PUBLIC_API_URL)
...
paths:
...
  /api/v1/tracks:
    get:
      tags: [tracks]
      summary: List or search published tracks
      parameters:
        - name: q
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/TrackStatus" }
        - name: creator_id
          in: query
          schema: { type: string, format: uuid }
        - name: sort
          in: query
          schema:
            type: string
            enum: [trending]
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TracksResponse"
  /api/v1/tracks/{id}:
    get:
      tags: [tracks]
      parameters:
        - $ref: "#/components/parameters/UuidId"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/creator/uploads/init:
    post:
      tags: [creator]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadInitInput"
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadInitResponse"
  /api/v1/creator/uploads/{trackID}/complete:
    post:
      tags: [creator]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: trackID
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                duration_ms: { type: integer }
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
  /api/v1/creator/uploads/{trackID}/retry:
    post:
      tags: [creator]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: trackID
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Re-queued
  /api/v1/creator/uploads/{trackID}/status:
    get:
      tags: [creator]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: trackID
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
  /api/v1/creator/tracks:
    get:
      tags: [creator]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TracksResponse"
  /api/v1/creator/tracks/{trackID}/cover/init:
    post:
      tags: [creator]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: trackID
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PresignUploadInput"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadInitResponse"
  /api/v1/creator/tracks/{trackID}/cover/complete:
    post:
      tags: [creator]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: trackID
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
...
  /api/v1/stream/playlist.m3u8:
    get:
      tags: [stream]
      summary: HLS master or variant playlist (requires pt token)
      parameters:
        - name: track_id
          in: query
          required: true
          schema: { type: string, format: uuid }
        - name: pt
          in: query
          required: true
          schema: { type: string }
        - name: path
          in: query
          schema: { type: string }
      responses:
        "200":
          description: MPEG-TS playlist
          content:
            application/vnd.apple.mpegurl:
              schema:
                type: string
  /api/v1/stream/hls.key:
    get:
      tags: [stream]
      summary: AES-128 HLS key (requires pt token)
      parameters:
        - name: track_id
          in: query
          required: true
          schema: { type: string, format: uuid }
        - name: pt
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Binary key
  /internal/admin/tracks:
    get:
      tags: [admin]
      security: [{ adminKey: [] }]
      parameters:
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/TrackStatus" }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/TrackWithCreator"
                  total: { type: integer }
  /internal/admin/tracks/{id}/approve:
    post:
      tags: [admin]
      security: [{ adminKey: [] }]
      parameters:
        - $ref: "#/components/parameters/UuidId"
      responses:
        "200":
          description: Published
  /internal/admin/tracks/{id}/reject:
    post:
      tags: [admin]
      security: [{ adminKey: [] }]
      parameters:
        - $ref: "#/components/parameters/UuidId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string }
      responses:
        "200":
          description: Rejected to draft
  /internal/admin/queue/dlq:
    get:
      tags: [admin]
      security: [{ adminKey: [] }]
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/DLQEntry"
  /internal/admin/queue/depths:
    get:
      tags: [admin]
      security: [{ adminKey: [] }]
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueDepths"
  /internal/admin/queue/dlq/retry:
    post:
      tags: [admin]
      security: [{ adminKey: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [track_id]
              properties:
                track_id: { type: string, format: uuid }
      responses:
        "200":
          description: Requeued
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminKey:
      type: apiKey
      in: header
      name: X-Gachify-Admin-Key
    seedKey:
      type: apiKey
      in: header
      name: X-Gachify-Seed-Key
  parameters:
    UuidId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    Limit:
      name: limit
      in: query
      schema: { type: integer, default: 20, maximum: 100 }
    Offset:
      name: offset
      in: query
      schema: { type: integer, default: 0, minimum: 0 }
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
        message: { type: string }
...
    TrackStatus:
      type: string
      enum:
        - draft
        - processing
        - pending_review
        - published
        - shadow_banned
        - removed
    Track:
      type: object
      properties:
        id: { type: string, format: uuid }
        creator_id: { type: string, format: uuid }
        title: { type: string }
        duration_ms: { type: integer }
        status: { $ref: "#/components/schemas/TrackStatus" }
        gachi_metadata: { type: object, additionalProperties: true }
        play_count: { type: integer }
        processing_error: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CreatorSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        handle: { type: string }
        display_name: { type: string }
    TrackWithCreator:
      allOf:
        - $ref: "#/components/schemas/Track"
        - type: object
          properties:
            creator:
              $ref: "#/components/schemas/CreatorSummary"
    TracksResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/TrackWithCreator"
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
        has_more: { type: boolean }
...
    UploadInitInput:
      type: object
      required: [title, filename, content_type]
      properties:
        title: { type: string }
        filename: { type: string }
        content_type: { type: string }
        duration_ms: { type: integer }
        gachi_metadata: { type: object, additionalProperties: true }
    UploadInitResponse:
      type: object
      properties:
        track_id: { type: string, format: uuid }
        upload_url: { type: string, format: uri }
        object_key: { type: string }
        expires_in_sec: { type: integer }
...
    QueueDepths:
      type: object
      properties:
        pending: { type: integer }
        processing: { type: integer }
        retry: { type: integer }
        dlq: { type: integer }
    DLQEntry:
      type: object
      properties:
        job:
          type: object
          properties:
            track_id: { type: string, format: uuid }
            job_id: { type: string, format: uuid }
        error: { type: string }
        failed_at: { type: string, format: date-time }